requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
langchain-core
langchain-openai
